requirements.txt
obs-websocket-py
flask
flask_cors
flask[async]
mss
pillow